MCUboot flash_area porting layer for PALTHREE and PALX
- Extract AT25SF641/W25Q512 ext-flash drivers to target_ext_flash.c/.h (shared by LittleFS and MCUboot)
- Refactor target_littlefs to use AT25SF641_*/W25Q512_* API
- Add mcuboot_flash_map.c: full flash_area_* impl for each board
- Add mcuboot_config.h per board
- Add stm32_f7xx_flash.h: shared F76xx internal flash sector geometry
- FindMCUboot.cmake: auto-discover mcuboot_flash_map.c from common/
- FindLITTLEFS.cmake: auto-discover target_ext_flash.c from common/
- MCUboot/CMakeLists.txt: activate as real STATIC lib (nf_mcuboot_port)
- Delete obsolete MCUboot/port/stm32/ stub files
