+++ MCUboot/CMakeLists.txt
@@ -14,41 +14,31 @@
 # via: if(NF_FEATURE_USE_MCUBOOT) add_subdirectory(MCUboot) endif()
 #
 # MCUboot upstream source is fetched via FindMCUboot.cmake (FetchContent).
-# The nanoFramework porting layer (port/stm32/ or port/esp32/) is compiled
-# together with MCUboot's bootutil library.
+# The nanoFramework porting layer is compiled as a static library (nf_mcuboot_port)
+# and linked by each board target.
 #
-# TODO(ifu-phase1): Wire up the full MCUboot bootloader build here.
-#   Step 1: add_library(mcuboot_port ...) with platform port sources.
-#   Step 2: target_include_directories for MCUboot upstream + nf port headers.
-#   Step 3: add_executable(mcuboot.elf) linking bootutil + mcuboot_port + BSP startup.
-#   Step 4: add post-build objcopy to produce mcuboot.bin / mcuboot.hex.
-#
-# For now this is a skeleton CMakeLists that documents the intended structure.
+# TODO(ifu-later): Wire up the full MCUboot bootloader executable here:
+#   - add_executable(mcuboot.elf) linking bootutil + nf_mcuboot_port + BSP startup
+#   - post-build objcopy to produce mcuboot.bin / mcuboot.hex
 
 cmake_minimum_required(VERSION 3.31)
 
-# MCUboot porting layer interface library.
-# Provides the include paths and port sources to any target that links MCUboot.
-add_library(nf_mcuboot_port INTERFACE)
-
-target_include_directories(nf_mcuboot_port INTERFACE
-    ${CMAKE_CURRENT_SOURCE_DIR}/include
-)
-
-# TODO(ifu-phase1): Replace INTERFACE with STATIC and add platform port sources:
+# MCUboot porting layer static library.
+# MCUboot_SOURCES and MCUboot_INCLUDE_DIRS are populated by FindMCUboot.cmake
+# (called via find_package(MCUboot REQUIRED) before add_subdirectory(MCUboot)).
 #
-# if(RTOS STREQUAL "CHIBIOS")
-#     target_sources(nf_mcuboot_port PRIVATE
-#         port/stm32/flash_map_stm32.c
-#     )
-#     target_include_directories(nf_mcuboot_port PUBLIC
-#         port/stm32
-#     )
-# elseif(RTOS STREQUAL "ESP32")
-#     target_sources(nf_mcuboot_port PRIVATE
-#         port/esp32/flash_map_esp32.c
-#     )
-#     target_include_directories(nf_mcuboot_port PUBLIC
-#         port/esp32
-#     )
-# endif()
+# mcuboot_flash_map.c is discovered by FindMCUboot.cmake using find_file() against
+# ${TARGET_BASE_LOCATION}/common — no board-specific path needed here.
+#
+# Each board's CMakeLists.txt is responsible for adding:
+#   - the external flash driver source (e.g. target_spi1_flash.c, target_qspi_flash.c)
+#   - board-specific include directories (e.g. _common/mcuboot for stm32_f7xx_flash.h)
+# via target_sources(nf_mcuboot_port PRIVATE ...) and
+#      target_include_directories(nf_mcuboot_port PUBLIC ...)
+
+add_library(nf_mcuboot_port STATIC ${MCUboot_SOURCES})
+
+target_include_directories(nf_mcuboot_port PUBLIC
+    ${MCUboot_INCLUDE_DIRS}
+    ${TARGET_BASE_LOCATION}
+)
